+++ nodes/out.js
@@ -144,14 +144,15 @@ module.exports = function(RED) {
                     //send data to API
                     var deviceMeta = node.server.getDevice(node.config.device);
                     if (deviceMeta !== undefined && deviceMeta && "device_id" in deviceMeta) {
+                        let url = 'http://' + node.server.ip + ':' + node.server.port + '/api/' + node.server.credentials.secured_apikey;
 						if (command == 'scene'){ // make a new URL for recalling the scene
 							var groupid = ((node.config.device).split('group_').join(''));
-                            var url = 'http://' + node.server.ip + ':' + node.server.port + '/api/' + node.server.apikey + '/groups/' + groupid + '/scenes/' + payload + '/recall';
+                            url += '/groups/' + groupid + '/scenes/' + payload + '/recall';
 						} else if ((/group_/g).test(node.config.device)) {
                             var groupid = ((node.config.device).split('group_').join(''));
-                            var url = 'http://' + node.server.ip + ':' + node.server.port + '/api/' + node.server.apikey + '/groups/' + groupid + '/action';
+                            url += '/groups/' + groupid + '/action';
                         } else {
-                            var url = 'http://' + node.server.ip + ':' + node.server.port + '/api/' + node.server.apikey + '/lights/' + deviceMeta.device_id + '/state';
+                            url += '/lights/' + deviceMeta.device_id + '/state';
                         }
                         var post = {};
 						if (node.commandType == 'object' || node.commandType == 'homekit') {